+++ test/sanity/ignore.txt
@@ -8297,6 +8297,10 @@ test/units/contrib/inventory/test_vmware_inventory.py future-import-boilerplate
 test/units/contrib/inventory/test_vmware_inventory.py metaclass-boilerplate
 test/units/contrib/inventory/test_vmware_inventory.py pylint:blacklisted-name
 test/units/executor/test_play_iterator.py pylint:blacklisted-name
+test/units/inventory/test_group.py future-import-boilerplate
+test/units/inventory/test_group.py metaclass-boilerplate
+test/units/inventory/test_host.py future-import-boilerplate
+test/units/inventory/test_host.py metaclass-boilerplate
 test/units/mock/path.py future-import-boilerplate
 test/units/mock/path.py metaclass-boilerplate
 test/units/mock/yaml_helper.py future-import-boilerplate
@@ -8361,6 +8365,8 @@ test/units/module_utils/network/ftd/test_fdm_swagger_with_real_data.py future-im
 test/units/module_utils/network/ftd/test_fdm_swagger_with_real_data.py metaclass-boilerplate
 test/units/module_utils/network/ftd/test_upsert_functionality.py future-import-boilerplate
 test/units/module_utils/network/ftd/test_upsert_functionality.py metaclass-boilerplate
+test/units/module_utils/network/netscaler/test_netscaler.py future-import-boilerplate
+test/units/module_utils/network/netscaler/test_netscaler.py metaclass-boilerplate
 test/units/module_utils/network/nso/test_nso.py metaclass-boilerplate
 test/units/module_utils/parsing/test_convert_bool.py future-import-boilerplate
 test/units/module_utils/postgresql/test_postgres.py future-import-boilerplate
@@ -8416,8 +8422,8 @@ test/units/modules/cloud/xenserver/FakeXenAPI.py metaclass-boilerplate
 test/units/modules/conftest.py future-import-boilerplate
 test/units/modules/conftest.py metaclass-boilerplate
 test/units/modules/files/test_copy.py future-import-boilerplate
-test/units/modules/messaging/rabbitmq/test_rabbimq_user.py future-import-boilerplate
-test/units/modules/messaging/rabbitmq/test_rabbimq_user.py metaclass-boilerplate
+test/units/modules/messaging/rabbitmq/test_rabbitmq_user.py future-import-boilerplate
+test/units/modules/messaging/rabbitmq/test_rabbitmq_user.py metaclass-boilerplate
 test/units/modules/monitoring/test_circonus_annotation.py future-import-boilerplate
 test/units/modules/monitoring/test_circonus_annotation.py metaclass-boilerplate
 test/units/modules/monitoring/test_icinga2_feature.py future-import-boilerplate
@@ -8468,8 +8474,6 @@ test/units/modules/network/netscaler/test_netscaler_lb_monitor.py future-import-
 test/units/modules/network/netscaler/test_netscaler_lb_monitor.py metaclass-boilerplate
 test/units/modules/network/netscaler/test_netscaler_lb_vserver.py future-import-boilerplate
 test/units/modules/network/netscaler/test_netscaler_lb_vserver.py metaclass-boilerplate
-test/units/modules/network/netscaler/test_netscaler_module_utils.py future-import-boilerplate
-test/units/modules/network/netscaler/test_netscaler_module_utils.py metaclass-boilerplate
 test/units/modules/network/netscaler/test_netscaler_nitro_request.py future-import-boilerplate
 test/units/modules/network/netscaler/test_netscaler_nitro_request.py metaclass-boilerplate
 test/units/modules/network/netscaler/test_netscaler_save_config.py future-import-boilerplate
@@ -8707,8 +8711,8 @@ test/units/modules/system/test_java_keystore.py metaclass-boilerplate
 test/units/modules/system/test_known_hosts.py future-import-boilerplate
 test/units/modules/system/test_known_hosts.py metaclass-boilerplate
 test/units/modules/system/test_known_hosts.py pylint:ansible-bad-function
-test/units/modules/system/test_linux_mountinfo.py future-import-boilerplate
-test/units/modules/system/test_linux_mountinfo.py metaclass-boilerplate
+test/units/modules/system/test_mount.py future-import-boilerplate
+test/units/modules/system/test_mount.py metaclass-boilerplate
 test/units/modules/system/test_pamd.py metaclass-boilerplate
 test/units/modules/system/test_parted.py future-import-boilerplate
 test/units/modules/system/test_systemd.py future-import-boilerplate
@@ -8735,10 +8739,6 @@ test/units/plugins/httpapi/test_ftd.py future-import-boilerplate
 test/units/plugins/httpapi/test_ftd.py metaclass-boilerplate
 test/units/plugins/inventory/test_constructed.py future-import-boilerplate
 test/units/plugins/inventory/test_constructed.py metaclass-boilerplate
-test/units/plugins/inventory/test_group.py future-import-boilerplate
-test/units/plugins/inventory/test_group.py metaclass-boilerplate
-test/units/plugins/inventory/test_host.py future-import-boilerplate
-test/units/plugins/inventory/test_host.py metaclass-boilerplate
 test/units/plugins/loader_fixtures/import_fixture.py future-import-boilerplate
 test/units/plugins/shell/test_cmd.py future-import-boilerplate
 test/units/plugins/shell/test_cmd.py metaclass-boilerplate
