+++ test/units/inventory/test_group.py
