+++ test/units/inventory/test_host.py
