+++ test/units/modules/messaging/rabbitmq/test_rabbitmq_user.py
