+++ test/units/modules/network/eric_eccli/test_eric_eccli_command.py
