+++ test/units/modules/network/onyx/test_onyx_ptp_global.py
