+++ test/units/plugins/filter/test_core.py
