+++ test/units/plugins/strategy/test_linear.py
