+++ test/units/plugins/strategy/test_strategy.py
